+++ src/Resources/contao/languages/de/modules.php
@@ -16,4 +16,4 @@ declare(strict_types=1);
  */
 $GLOBALS['TL_LANG']['MOD']['synapsis'] = 'Synapsis-Forum';
 $GLOBALS['TL_LANG']['MOD']['synapsis_forum'] = array('Forum', 'Die Forenstruktur mit Kategorien, Foren, Themen und Beiträgen verwalten.');
-$GLOBALS['TL_LANG']['MOD']['synapsis_csv'] = array('CSV Import / Export', 'Eine Forenstruktur als CSV exportieren oder wiederherstellen.');
+$GLOBALS['TL_LANG']['MOD']['synapsis_csv'] = array('CSV Import', 'Eine Forenstruktur aus zwei CSV-Dateien (Kategorien/Foren und Themen/Beiträge) importieren, z. B. aus phpBB.');